Added Dolphin + Adjusted Net
Added in a Dolphin Object to try testing Co-op gameplay.
Currently only copies the Boat Object around on screen with the same controls.

Also adjusted the 'Net' object as it was spawning above the boat within the code and so couldn't be seen on screen.
Now it'll appear but there is an issue with it's position updating strangely.